pyproject.toml
[project]
name = "country-info-agent-cloudeagle"
version = "0.1.0"
description = "Country Information AI Agent using LangGraph + Chainlit"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.4.0",
    "httpx>=0.28.1",
    "langchain>=0.3.0",
    "langchain-openrouter>=0.2.0",
    "langgraph>=0.3.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.1.4",
]
